add documentation for new public classes
