+++ dart/lib/src/exception_cause.dart
@@ -1,3 +1,4 @@
+/// Holds inner exception and stackTrace combinations contained in other exceptions
 class ExceptionCause {
   ExceptionCause(this.exception, this.stackTrace);
