+++ dart/lib/src/recursive_exception_cause_extractor.dart
@@ -1,0 +1,47 @@
+import 'sentry_options.dart';
+import 'throwable_mechanism.dart';
+import 'exception_cause.dart';
+import 'protocol.dart';
+
+/// Extracts inner exceptions recursively
+class RecursiveExceptionCauseExtractor {
+  RecursiveExceptionCauseExtractor(this._options);
+
+  final SentryOptions _options;
+
+  List<ExceptionCause> flatten(exception, stackTrace) {
+    final allExceptionCauses = <ExceptionCause>[];
+    final circularityDetector = <dynamic>{};
+
+    var currentException = exception;
+    ExceptionCause? currentExceptionCause =
+        ExceptionCause(exception, stackTrace);
+
+    while (currentException != null &&
+        currentExceptionCause != null &&
+        circularityDetector.add(currentException)) {
+      allExceptionCauses.add(currentExceptionCause);
+
+      final extractionSourceSource = currentException is ThrowableMechanism
+          ? currentException.throwable
+          : currentException;
+
+      final extractor =
+          _options.exceptionCauseExtractor(extractionSourceSource.runtimeType);
+
+      try {
+        currentExceptionCause = extractor?.cause(extractionSourceSource);
+        currentException = currentExceptionCause?.exception;
+      } catch (exception, stackTrace) {
+        _options.logger(
+          SentryLevel.error,
+          'An exception occurred while extracting  exception cause',
+          exception: exception,
+          stackTrace: stackTrace,
+        );
+        break;
+      }
+    }
+    return allExceptionCauses;
+  }
+}
